+++ source/patterns/@aws-solutions-constructs/aws-dynamodb-stream-lambda/test/integ.no-arguments.ts
@@ -15,10 +15,12 @@
 import { App, Stack } from "@aws-cdk/core";
 import { DynamoDBStreamToLambdaProps, DynamoDBStreamToLambda } from "../lib";
 import * as lambda from '@aws-cdk/aws-lambda';
+import { generateIntegStackName } from '@aws-solutions-constructs/core';
+
 const app = new App();
 
 // Empty arguments
-const stack = new Stack(app, 'test-dynamodb-stream-lambda-stack');
+const stack = new Stack(app, generateIntegStackName(__filename));
 
 const props: DynamoDBStreamToLambdaProps = {
   lambdaFunctionProps: {