+++ source/patterns/@aws-solutions-constructs/aws-kinesisstreams-gluejob/test/integ.no-arguments.ts
@@ -16,10 +16,11 @@ import { CfnTable } from '@aws-cdk/aws-glue';
 import { Asset } from '@aws-cdk/aws-s3-assets';
 import { App, Stack } from '@aws-cdk/core';
 import { KinesisstreamsToGluejob } from '../lib';
+import { generateIntegStackName } from '@aws-solutions-constructs/core';
 
 // Setup
 const app = new App();
-const stack = new Stack(app, 'test-kinesisstream-gluejob');
+const stack = new Stack(app, generateIntegStackName(__filename));
 stack.templateOptions.description = 'Integration Test for aws-kinesisstream-gluejob';
 
 const fieldSchema: CfnTable.ColumnProperty [] = [{