+++ source/patterns/@aws-solutions-constructs/aws-sns-sqs/test/integ.no-arguments.ts
@@ -15,10 +15,11 @@
 import { App, Stack } from "@aws-cdk/core";
 import { SnsToSqs, SnsToSqsProps } from "../lib";
 import * as iam from '@aws-cdk/aws-iam';
+import { generateIntegStackName } from '@aws-solutions-constructs/core';
 
 // Setup
 const app = new App();
-const stack = new Stack(app, 'test-sns-sqs-stack');
+const stack = new Stack(app, generateIntegStackName(__filename));
 stack.templateOptions.description = 'Integration Test for aws-sns-sqs';
 
 // Definitions