+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.106.1](https://github.com/awslabs/aws-solutions-constructs/compare/v1.106.0...v1.106.1) (2021-06-03)
+
+### Changed
+
+* Upgraded all patterns to CDK v1.106.1
+
+## [1.106.0](https://github.com/awslabs/aws-solutions-constructs/compare/v1.105.0...v1.106.0) (2021-06-02)
+
+### Changed
+
+* Upgraded all patterns to CDK v1.106.0
+* Fixed stack name duplication for ALL integ tests [#183](https://github.com/awslabs/aws-solutions-constructs/issues/183)
+* Fixed aws-cloudfront-mediastore integ tests failure [#194](https://github.com/awslabs/aws-solutions-constructs/issues/194)
+
+## [1.105.0](https://github.com/awslabs/aws-solutions-constructs/compare/v1.104.0...v1.105.0) (2021-05-28)
+
+### Changed
+
+* Upgraded all patterns to CDK v1.105.0
+* aws-s3-step-function now triggers for Mulitpart uploads and Copy Object [#198](https://github.com/awslabs/aws-solutions-constructs/pull/198)
+* Change S3 protocol in glue to use s3a:// [#197](https://github.com/awslabs/aws-solutions-constructs/pull/197)
+* Change policy name in for glue so multiple constructs can be included in a single stack [#197](https://github.com/awslabs/aws-solutions-constructs/pull/197)
+
+### ⚠ BREAKING CHANGES
+
+* Existing testing snapshots for aws-s3-step-function and aws-kinesisstreams-gluejob may need to be refreshed.
+
 ## [1.104.0](https://github.com/awslabs/aws-solutions-constructs/compare/v1.103.0...v1.104.0) (2021-05-21)
 
 ### Changed